Versionamento controllers, dtos e services.

+++ backend/bookstore/src/main/java/com/poletto/bookstore/services/v2/BookService.java
@@ -11,17 +11,15 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.poletto.bookstore.controllers.v1.BookControllerV1;
+import com.poletto.bookstore.controllers.v2.BookController;
 import com.poletto.bookstore.converter.custom.BookMapper;
-import com.poletto.bookstore.dto.v1.BookDTOv1;
-import com.poletto.bookstore.dto.v1.CategoryDTOv1;
 import com.poletto.bookstore.dto.v2.BookDTOv2;
+import com.poletto.bookstore.dto.v2.CategoryDTOv2;
 import com.poletto.bookstore.entities.Book;
 import com.poletto.bookstore.entities.Category;
 import com.poletto.bookstore.entities.enums.BookStatus;
@@ -33,10 +31,10 @@ import com.poletto.bookstore.repositories.CategoryRepository;
 
 import jakarta.persistence.EntityNotFoundException;
 
-@Service
-public class BookServiceV2 {
-	
-	private static final Logger logger = LoggerFactory.getLogger(BookServiceV2.class);
+@Service("BookServiceV2")
+public class BookService {
+
+	private static final Logger logger = LoggerFactory.getLogger(BookService.class);
 
 	@Autowired
 	private BookRepository bookRepository;
@@ -47,6 +45,8 @@ public class BookServiceV2 {
 	@Autowired
 	private AuthorRepository authorRepository;
 	
+	//TODO adicionar pesquisa por autor
+
 	@Transactional(readOnly = true)
 	public Page<BookDTOv2> findAllPaged(Pageable pageable, Long categoryId, String name, String booked) {
 
@@ -69,79 +69,103 @@ public class BookServiceV2 {
 
 		}
 
-		var bookPage = bookRepository.findPaged(categories, name, booked, pageable);
+		Page<Book> bookPage = bookRepository.findPaged(categories, name, booked, pageable);
 
 		bookRepository.findProductsWithCategories(bookPage.getContent());
-		
-		logger.info("Resource BOOK page found: " + "PAGE NUMBER [" + bookPage.getNumber() + "] - CONTENT: " + bookPage.getContent());
-		
+
+		logger.info("Resource BOOK page found: " + "PAGE NUMBER [" + bookPage.getNumber() + "] - CONTENT: "
+				+ bookPage.getContent());
+
 		Page<BookDTOv2> dtos = bookPage.map(x -> BookMapper.convertEntityToDtoV2(x));
-		
-		dtos.stream().forEach(x -> x.add(linkTo(methodOn(BookControllerV1.class).findById(x.getId())).withSelfRel()));
+
+		dtos.stream().forEach(x -> x
+				.add(linkTo(methodOn(BookController.class).findById(x.getId())).withSelfRel().withType("GET"))
+				.add(linkTo(methodOn(BookController.class).delete(x.getId())).withRel("delete").withType("DELETE"))
+				.add(linkTo(methodOn(BookController.class).update(x.getId(), x)).withRel("update").withType("PUT")));
 
 		return dtos;
 
 	}
 
 	@Transactional(readOnly = true)
-	public BookDTOv1 findById(Long id) {
-		
+	public BookDTOv2 findById(Long id) {
+
 		Optional<Book> obj = bookRepository.findById(id);
-		
+
 		Book book = obj.orElseThrow(() -> new ResourceNotFoundException("Resource BOOK not found. ID " + id));
-		
+
 		logger.info("Resource BOOK found: " + book.toString());
-		
-		return BookMapper.convertEntityToDto(book);
-		
+
+		BookDTOv2 dto = BookMapper.convertEntityToDtoV2(book);
+
+		dto.add(linkTo(methodOn(BookController.class).findById(dto.getId())).withSelfRel().withType("GET"));
+		dto.add(linkTo(methodOn(BookController.class).delete(dto.getId())).withRel("delete").withType("DELETE"));
+		dto.add(linkTo(methodOn(BookController.class).update(dto.getId(), dto)).withRel("update").withType("PUT"));
+
+		return dto;
+
 	}
 
+	// TODO n esta retornando categorias no response de book.insert
+
 	@Transactional
-	public BookDTOv2 insert(BookDTOv2 dto) {
-		
+	public BookDTOv2 insert(BookDTOv2 dto) {	
+
 		dto.setStatus(BookStatus.AVAILABLE);
-		
+
 		dto.setAuthor(authorRepository.findById(dto.getAuthor().getId()).orElseThrow(() -> new ResourceNotFoundException("Resource AUTHOR not found. ID " + dto.getAuthor().getId())));
 
 		Book entity = BookMapper.convertDtoToEntityV2(dto);
-		
+
 		entity.getCategories().clear();
-		
-		for (CategoryDTOv1 categoryDTO : dto.getCategories()) {
+
+		for (CategoryDTOv2 categoryDTO : dto.getCategories()) {
 			try {
-			Category category = categoryRepository.getReferenceById(categoryDTO.getId());
-			entity.getCategories().add(category);
+				Category category = categoryRepository.getReferenceById(categoryDTO.getId());
+				entity.getCategories().add(category);
 			} catch (EntityNotFoundException ex) {
 				throw new ResourceNotFoundException("Resource CATEGORY not found. ID " + categoryDTO.getId());
 			}
 		}
 
 		entity = bookRepository.save(entity);
-		
-		logger.info("Resource BOOK saved: " + entity.toString());
 
-		return BookMapper.convertEntityToDtoV2(entity);
+		logger.info("Resource BOOK saved: " + entity.toString());
+		
+		BookDTOv2 newDto = BookMapper.convertEntityToDtoV2(entity);
+
+		newDto.add(linkTo(methodOn(BookController.class).findById(newDto.getId())).withSelfRel().withType("GET"));
+		newDto.add(linkTo(methodOn(BookController.class).delete(newDto.getId())).withRel("delete").withType("DELETE"));
+		newDto.add(linkTo(methodOn(BookController.class).update(newDto.getId(), newDto)).withRel("update").withType("PUT"));
+		
+		return newDto;
 
 	}
 
 	@Transactional
-	public BookDTOv1 update(Long id, BookDTOv1 dto) {
+	public BookDTOv2 update(Long id, BookDTOv2 dto) {
 
 		try {
-			
+
 			Book entity = bookRepository.getReferenceById(id);
-			
+
 			dto.setStatus(entity.getStatus());
-			
+
 			dto.setId(entity.getId());
 
-			entity = BookMapper.convertDtoToEntity(dto);
+			entity = BookMapper.convertDtoToEntityV2(dto);
 
 			entity = bookRepository.save(entity);
-			
+
 			logger.info("Resource BOOK updated: " + entity.toString());
 
-			return BookMapper.convertEntityToDto(entity);
+			dto = BookMapper.convertEntityToDtoV2(entity);
+
+			dto.add(linkTo(methodOn(BookController.class).findById(dto.getId())).withSelfRel().withType("GET"));
+			dto.add(linkTo(methodOn(BookController.class).delete(dto.getId())).withRel("delete").withType("DELETE"));
+			dto.add(linkTo(methodOn(BookController.class).update(dto.getId(), dto)).withRel("update").withType("PUT"));
+
+			return dto;
 
 		} catch (EntityNotFoundException e) {
 
@@ -155,13 +179,13 @@ public class BookServiceV2 {
 
 		try {
 
-			bookRepository.deleteById(id);
-			
+			if (bookRepository.existsById(id)) {
+				bookRepository.deleteById(id);
+			} else {
+				throw new ResourceNotFoundException("Resource BOOK not found. ID " + id);
+			}
+
 			logger.info("Resource BOOK deleted: ID " + id);
-
-		} catch (EmptyResultDataAccessException e) {
-
-			throw new ResourceNotFoundException("Resource BOOK not found. ID " + id);
 
 		} catch (DataIntegrityViolationException e) {
